requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
donfig==0.8.1.post1
fonttools==4.61.1
fsspec==2025.12.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.8.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.14.0
netCDF4==1.7.4
numcodecs==0.16.5
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
-e git+ssh://git@github.com/rmower90/snow-ops-streamlit.git@9b9829896efccee93bc5a553ca9dc00b4aa72cde#egg=snow_ops
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
xarray==2025.12.0
zarr==3.1.5
